- hbm2ddl.auto set to update - refactoring to separate common and postgres-specific parts of API
