+++ bdse-kvnode/src/main/java/ru/csc/bdse/Application.java
@@ -4,7 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import ru.csc.bdse.kv.KeyValueApi;
-import ru.csc.bdse.kv.db.PersistentKeyValueApi;
+import ru.csc.bdse.kv.db.PostgresPersistentKeyValueApi;
 import ru.csc.bdse.util.Env;
 
 import java.util.UUID;
@@ -14,6 +14,13 @@ public class Application {
 
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
+
+        final KeyValueApi node = new Application().node();
+//        node.put("SomeKey", "SomeValue".getBytes());
+//        node.put("OneMoreKey", "OneMoreValue".getBytes());
+
+        System.out.println(new String(node.get("SomeKey").get()));
+        System.out.println(new String(node.get("OneMoreKey").get()));
     }
 
     private static String randomNodeName() {
@@ -23,6 +30,6 @@ public class Application {
     @Bean
     KeyValueApi node() {
         String nodeName = Env.get(Env.KVNODE_NAME).orElseGet(Application::randomNodeName);
-        return new PersistentKeyValueApi(nodeName);
+        return new PostgresPersistentKeyValueApi(nodeName);
     }
 }
