+++ bdse-kvnode/src/main/java/ru/csc/bdse/kv/db/PersistentKeyValueApi.java
@@ -4,12 +4,7 @@ import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
-import org.hibernate.cfg.Configuration;
-import org.jetbrains.annotations.NotNull;
 import ru.csc.bdse.kv.KeyValueApi;
-import ru.csc.bdse.kv.NodeAction;
-import ru.csc.bdse.kv.NodeInfo;
-import ru.csc.bdse.kv.NodeStatus;
 
 import java.util.*;
 import java.util.function.Function;
@@ -18,23 +13,9 @@ import java.util.function.Function;
  * Implementation of persistent key-value storage.
  *
  */
-public final class PersistentKeyValueApi implements KeyValueApi {
+public abstract class PersistentKeyValueApi implements KeyValueApi {
 
-    @NotNull private final SessionFactory factory;
-    @NotNull private final NodeInfo state;
-
-    public PersistentKeyValueApi(@NotNull String name) {
-        this.state = new NodeInfo(name, NodeStatus.UP);
-
-        try {
-            factory = new Configuration().configure()
-                    .addAnnotatedClass(Entity.class)
-                    .buildSessionFactory();
-        } catch (Throwable ex) {
-            System.err.println("Failed to create sessionFactory object." + ex);
-            throw new ExceptionInInitializerError(ex);
-        }
-    }
+    protected SessionFactory factory;
 
     private <T> T runQuery(Function<Session, T> fun) {
         Transaction tx = null;
@@ -93,18 +74,6 @@ public final class PersistentKeyValueApi implements KeyValueApi {
             session.delete(entity);
             return null;
         });
-    }
-
-    @Override
-    public Set<NodeInfo> getInfo() {
-        return Collections.singleton(state);
-    }
-
-    @Override
-    public void action(String node, NodeAction action) {
-        if (node.equals(state.getName())) {
-            state.setStatus(action);
-        }
     }
 
 }
