+++ bdse-kvnode/src/main/resources/hibernate_postgres.cfg.xml
@@ -24,7 +24,7 @@
         </property>
 
         <property name="connection_pool_size">1</property>
-        <property name="hbm2ddl.auto">none</property> <!-- something like update, but for some reason throws exc -->
+        <property name="hbm2ddl.auto">update</property>
         <property name="show_sql">true</property>
 
     </session-factory>
